+++ chal/hashmap.h
@@ -2,17 +2,17 @@
 //#include <libdex/sha1.h>
 
 struct HashEntry {
-    size_t key;
-    size_t data;
+    uint64_t key;
+    uint64_t data;
     HashEntry *next;
 };
 
 struct HashMap {
-    size_t size;
+    uint64_t size;
     HashEntry *root;
 };
 
-size_t lookup(HashMap *map, size_t key) {
+uint64_t lookup(HashMap *map, uint64_t key) {
     HashEntry *e = NULL;
     for (e = map->root; e != NULL; e = e->next) {
         if (e->key == key)
@@ -21,7 +21,7 @@ size_t lookup(HashMap *map, size_t key) {
     return 0;
 }
 
-void insert(HashMap *map, size_t key, size_t data) {
+void insert(HashMap *map, uint64_t key, uint64_t data) {
     HashEntry *e = (HashEntry*)malloc(sizeof(HashEntry));
     e->key = key;
     e->data = data;
