+++ chal/linkframe.h
@@ -1,9 +1,10 @@
 
 #include <stdlib.h>
+#include <stdint.h>
 
 struct LinkEntry {
-    size_t regno;
-    size_t regval;
+    uint64_t regno;
+    uint64_t regval;
     LinkEntry *next;
 };
 
@@ -12,7 +13,7 @@ struct LinkEntry {
  *  - Act in a stack to handle nested subroutine invoke/call
  */
 struct LinkFrame {
-    //size_t regs[65537]; // TOO big ~512 kb, use linklist
+    //uint64_t regs[65537]; // TOO big ~512 kb, use linklist
     LinkEntry *reg;
     void *pc;
     LinkFrame *next;
@@ -20,10 +21,10 @@ struct LinkFrame {
 
 // Pop the current frame from the frame stack
 // Return the pc of current frame
-void *restoreFrame(LinkFrame **curframe, size_t *regset);
+void *restoreFrame(LinkFrame **curframe, uint64_t *regset);
 
 // Create a clean slate for the subroutine
 // - Stash the regs to the frame stack
 // - NULL all the regs
-void stashFrame(size_t *regset, void *pc, LinkFrame **curframe);
+void stashFrame(uint64_t *regset, void *pc, LinkFrame **curframe);
